requirements.txt
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
mlflow>=2.9.0
azure-ai-ml>=1.19.0
azure-identity>=1.16.0
python-dotenv>=1.0.0
joblib>=1.3.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit